+++ src/contexts/Pools/BondedPools.tsx
@@ -1,0 +1,212 @@
+// Copyright 2022 @paritytech/polkadot-staking-dashboard authors & contributors
+// SPDX-License-Identifier: Apache-2.0
+
+import BN from 'bn.js';
+import React, { useState, useEffect, useRef } from 'react';
+import { bnToU8a, u8aConcat } from '@polkadot/util';
+import { APIContextInterface } from 'types/api';
+import { BondedPoolsContextState, PoolsConfigContextState } from 'types/pools';
+import { EMPTY_H256, MOD_PREFIX, U32_OPTS } from 'consts';
+import { useApi } from '../Api';
+import { usePoolsConfig } from './Config';
+import { setStateWithRef } from '../../Utils';
+
+export const BondedPoolsContext =
+  React.createContext<BondedPoolsContextState | null>(null);
+
+export const useBondedPools = () => React.useContext(BondedPoolsContext);
+
+export const BondedPoolsProvider = ({
+  children,
+}: {
+  children: React.ReactNode;
+}) => {
+  const { api, network, isReady, consts } = useApi() as APIContextInterface;
+  const { enabled } = usePoolsConfig() as PoolsConfigContextState;
+  const { poolsPalletId } = consts;
+
+  // stores the meta data batches for pool lists
+  const [poolMetaBatches, setPoolMetaBatch]: any = useState({});
+  const poolMetaBatchesRef = useRef(poolMetaBatches);
+
+  // stores the meta batch subscriptions for pool lists
+  const [poolSubs, setPoolSubs]: any = useState({});
+  const poolSubsRef = useRef(poolSubs);
+
+  // store bonded pools
+  const [bondedPools, setBondedPools]: any = useState([]);
+
+  useEffect(() => {
+    if (isReady && enabled) {
+      fetchBondedPools();
+    }
+    return () => {
+      unsubscribe();
+    };
+  }, [network, isReady, enabled]);
+
+  const unsubscribe = () => {
+    Object.values(poolSubsRef.current).map((batch: any, i: number) => {
+      return Object.entries(batch).map(([k, v]: any) => {
+        return v();
+      });
+    });
+    setBondedPools([]);
+  };
+
+  // fetch all bonded pool entries
+  const fetchBondedPools = async () => {
+    if (!api) return;
+
+    const _exposures = await api.query.nominationPools.bondedPools.entries();
+    // humanise exposures to send to worker
+    const exposures = _exposures.map(([_keys, _val]: any) => {
+      const id = _keys.toHuman()[0];
+      const pool = _val.toHuman();
+      return getPoolWithAddresses(id, pool);
+    });
+
+    setBondedPools(exposures);
+  };
+
+  /*
+    Fetches a new batch of pool metadata.
+    Fetches the metadata of a pool that we assume to be a string.
+    structure:
+    {
+      key: {
+        [
+          {
+          metadata: [],
+        }
+      ]
+    },
+  };
+  */
+  const fetchPoolsMetaBatch = async (key: string, p: any, refetch = false) => {
+    if (!isReady || !api) {
+      return;
+    }
+    if (!p.length) {
+      return;
+    }
+    if (!refetch) {
+      // if already exists, do not re-fetch
+      if (poolMetaBatchesRef.current[key] !== undefined) {
+        return;
+      }
+    } else {
+      // tidy up if existing batch exists
+      delete poolMetaBatches[key];
+      delete poolMetaBatchesRef.current[key];
+
+      if (poolSubsRef.current[key] !== undefined) {
+        for (const unsub of poolSubsRef.current[key]) {
+          unsub();
+        }
+      }
+    }
+
+    const ids = [];
+    for (const _p of p) {
+      ids.push(Number(_p.id));
+    }
+
+    // store batch ids
+    const batchesUpdated = Object.assign(poolMetaBatchesRef.current);
+    batchesUpdated[key] = {};
+    batchesUpdated[key].ids = ids;
+    setStateWithRef(
+      { ...batchesUpdated },
+      setPoolMetaBatch,
+      poolMetaBatchesRef
+    );
+
+    const subscribeToMetadata = async (id: any) => {
+      const unsub = await api.query.nominationPools.metadata.multi(
+        id,
+        (_metadata: any) => {
+          const metadata = [];
+          for (let i = 0; i < _metadata.length; i++) {
+            metadata.push(_metadata[i].toHuman());
+          }
+          const _batchesUpdated = Object.assign(poolMetaBatchesRef.current);
+          _batchesUpdated[key].metadata = metadata;
+
+          setStateWithRef(
+            { ..._batchesUpdated },
+            setPoolMetaBatch,
+            poolMetaBatchesRef
+          );
+        }
+      );
+      return unsub;
+    };
+
+    // initiate subscriptions
+    await Promise.all([subscribeToMetadata(ids)]).then((unsubs: any) => {
+      addMetaBatchUnsubs(key, unsubs);
+    });
+  };
+
+  /*
+   * Helper: to add mataBatch unsubs by key.
+   */
+  const addMetaBatchUnsubs = (key: string, unsubs: any) => {
+    const _unsubs = poolSubsRef.current;
+    const _keyUnsubs = _unsubs[key] ?? [];
+
+    _keyUnsubs.push(...unsubs);
+    _unsubs[key] = _keyUnsubs;
+    setStateWithRef(_unsubs, setPoolSubs, poolSubsRef);
+  };
+
+  /*
+   *  Helper: to add addresses to pool record.
+   */
+  const getPoolWithAddresses = (id: number, pool: any) => {
+    return {
+      ...pool,
+      id,
+      addresses: createAccounts(id),
+    };
+  };
+
+  // Helper: generates pool stash and reward accounts. assumes poolsPalletId is synced.
+  const createAccounts = (poolId: number): any => {
+    const poolIdBN = new BN(poolId);
+    return {
+      stash: createAccount(poolIdBN, 0),
+      reward: createAccount(poolIdBN, 1),
+    };
+  };
+
+  const createAccount = (poolId: BN, index: number): string => {
+    if (!api) return '';
+    return api.registry
+      .createType(
+        'AccountId32',
+        u8aConcat(
+          MOD_PREFIX,
+          poolsPalletId,
+          new Uint8Array([index]),
+          bnToU8a(poolId, U32_OPTS),
+          EMPTY_H256
+        )
+      )
+      .toString();
+  };
+
+  return (
+    <BondedPoolsContext.Provider
+      value={{
+        fetchPoolsMetaBatch,
+        createAccounts,
+        bondedPools,
+        meta: poolMetaBatchesRef.current,
+      }}
+    >
+      {children}
+    </BondedPoolsContext.Provider>
+  );
+};
